[SHARE-848][Feature]Move 'force_text' to utils.py * Move force_text from oai, clinicaltrials, mods, and datacite to utils * Use one version for all sources * Remove text_list and replace with force_text
