+++ share/transform/chain/utils.py
@@ -1,3 +1,8 @@
+import logging
+
+logger = logging.getLogger(__name__)
+
+
 def format_address(address1='', address2='', city='', state_or_province='', postal_code='', country=''):
     if address1 and address2 and city and state_or_province and postal_code and country:
         return '{}\n{}\n{}, {} {}\n{}'.format(address1, address2, city, state_or_province, postal_code, country)
@@ -36,3 +41,51 @@ def format_address(address1='', address2='', city='', state_or_province='', post
         return '{}, {}'.format(city, state_or_province)
 
     return address1
+
+
+def text_list(self, data):
+    text_list = []
+    if isinstance(data, list):
+        for item in data:
+            if isinstance(item, dict):
+                if '#text' in item:
+                    text_list.append(item['#text'])
+                    continue
+            elif isinstance(item, str):
+                text_list.append(item)
+                continue
+            logger.warning('#text is not in {} and it is not a string'.format(item))
+        return text_list
+    else:
+        raise Exception('{} is not a list.'.format(data))
+
+
+def force_text(data):
+    if isinstance(data, dict):
+        if '#text' in data:
+            return data['#text']
+        raise Exception('#text is not in {}'.format(data))
+
+    if isinstance(data, str):
+        return data
+
+    if isinstance(data, list):
+        text_list = []
+        for datum in (data or []):
+            if datum is None:
+                continue
+            if isinstance(datum, dict):
+                if '#text' not in datum:
+                    logger.warning('Skipping %s, no #text key exists', datum)
+                    continue
+                text_list.append(datum['#text'])
+            elif isinstance(datum, str):
+                text_list.append(datum)
+            else:
+                raise Exception(datum)
+        return text_list
+
+    if data is None:
+        return ''
+
+    raise TypeError(data)
